pyproject.toml
[tool.poetry]
name = "vibematching"
version = "0.1.0"
description = "VibeMatching backend"
authors = ["VibeMatching Team"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
sqlalchemy = "^2.0.35"
asyncpg = "^0.29.0"
psycopg = { extras = ["binary"], version = "^3.2.6" }
alembic = "^1.13.2"
pydantic-settings = "^2.6.1"
redis = "^5.0.8"
aio-pika = "^9.4.3"
minio = "^7.2.9"
python-multipart = "^0.0.9"
aiogram = "^3.13.1"
httpx = "^0.27.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"
ruff = "^0.6.8"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
